added tests for MainWindowViewModel
