+++ Assignment9/Assignment9.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31025.194
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assignment9", "Assignment9\Assignment9.csproj", "{B6B2BE3C-12E8-46BD-9C73-7508A5ECACE2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Assignment9", "Assignment9\Assignment9.csproj", "{B6B2BE3C-12E8-46BD-9C73-7508A5ECACE2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Assignment9.Tests", "Assignment9.Tests\Assignment9.Tests.csproj", "{EE6A5AA5-A9C7-44A5-8BC9-8997A491E4D2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{B6B2BE3C-12E8-46BD-9C73-7508A5ECACE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B6B2BE3C-12E8-46BD-9C73-7508A5ECACE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B6B2BE3C-12E8-46BD-9C73-7508A5ECACE2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE6A5AA5-A9C7-44A5-8BC9-8997A491E4D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE6A5AA5-A9C7-44A5-8BC9-8997A491E4D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE6A5AA5-A9C7-44A5-8BC9-8997A491E4D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE6A5AA5-A9C7-44A5-8BC9-8997A491E4D2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
